pyproject.toml
[project]
name = "netbox-kea-ng"
version = "1.1.1"
description = "NetBox plugin for the Kea DHCP server (fork of netbox-kea)"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
license-files = ["LICENSE", "LICENSES/*.txt"]
authors = [
  {name = "Devon Mar", email = "devon-mar@users.noreply.github.com"},
  {name = "Marcin Zieba"},
]
keywords = [
  "netbox",
  "plugin",
  "kea",
  "dhcp",
  "networking",
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Framework :: Django",
  "Environment :: Plugins",
  "Topic :: System :: Networking",
]
dependencies = [
  "requests>=2.27.0,<3.0.0",
  "netaddr>=0.8,<2.0.0",
]

[project.urls]
Homepage = "https://github.com/marcinpsk/netbox-kea"
Repository = "https://github.com/marcinpsk/netbox-kea"
Issues = "https://github.com/marcinpsk/netbox-kea/issues"
Changelog = "https://github.com/marcinpsk/netbox-kea/blob/main/CHANGELOG.md"

[dependency-groups]
dev = [
  "types-requests>=2.27.0,<3.0.0",
  "mypy>=1.14.0,<1.20.0",
  "pytest>=8.0.0,<10.0.0",
  "pytest-cov>=6.0",
  "pytest-django>=4.9.0,<5.0.0",
  "pytest-playwright>=0.6.0,<0.8.0",
  "pynetbox>=7.3.0,<7.7.0",
  "django-stubs[compatible-mypy]>=5.0.0,<6.0.0",
  "ruff>=0.8.0",
  "reuse",
  "pre-commit",
  "python-semantic-release",
  "tblib>=3.0.0,<4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = [
  "netbox_kea",
]

[tool.hatch.build.targets.wheel]
packages = ["netbox_kea"]
exclude = ["netbox_kea/tests"]

[tool.ruff]
line-length = 120
exclude = [
  "netbox_kea/migrations",
]

[tool.ruff.lint]
select = [
  "C4",    # flake8-comprehensions
  "C901",  # mccabe complexity
  "D",     # pydocstyle
  "E",
  "F",
  "I",
  "PERF",  # perflint
  "RET",   # flake8-return
  "UP",
  "W",
]
ignore = [
  "D100",   # missing docstring in public module
  "D104",   # missing docstring in public package
  "D105",   # missing docstring in magic method
  "D106",   # missing docstring in public nested class (Django Meta, etc.)
  "D107",   # missing docstring in __init__
  "D203",   # 1 blank line before class docstring (conflicts with D211)
  "D213",   # multi-line summary second line (conflicts with D212)
  "E501",   # line too long (handled by formatter)
  # https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
  "W191",
  "E111",
  "E114",
  "E117",
  "D206",
  "D300",
  "COM812",
  "COM819",
  "ISC001",
  "ISC002",
]

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"*/migrations/*.py" = ["D"]
"*/tests/*.py" = ["D"]
"tests/*.py" = ["D"]
"e2e/*.py" = ["D"]
".devcontainer/**/*.py" = ["D", "C4"]
"netbox_kea/forms.py" = ["C901"]  # upstream complex validation logic

[tool.pytest.ini_options]
pythonpath = ["/opt/netbox/netbox"]
DJANGO_SETTINGS_MODULE = "netbox.settings"
testpaths = ["netbox_kea/tests"]
addopts = "--ignore=tests/docker"

[tool.coverage.run]
source = ["netbox_kea"]
omit = ["*/migrations/*"]

[tool.coverage.report]
show_missing = true

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
commit_message = "chore(release): {version}\n\nAutomatically generated by python-semantic-release"
build_command = "uv build"

[tool.semantic_release.branches.main]
match = "^main$"
prerelease = false

[tool.semantic_release.publish]
upload_to_vcs_release = true

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"

[tool.semantic_release.commit_parser_options]
allowed_tags = ["feat", "fix", "docs", "style", "refactor", "perf", "test", "build", "ci", "chore", "revert"]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]
parse_squash_commits = false
